Separate google login button as component.
