+++ src/App.tsx
@@ -7,20 +7,23 @@ import { Router } from "./Router";
 import { Header } from "./components/Header";
 import { Sidebar } from "./components/Sidebar";
 import { AppThemeProvider } from "./components/AppThemeProvider";
+import { GoogleAuthProvider } from "./components/GoogleAuthProvider";
 
 function App() {
   return (
     <Provider store={store}>
       <BrowserRouter>
         <AppThemeProvider>
-          <Box sx={{ display: " flex", height: "100vh" }}>
-            <Header />
-            <Sidebar />
-            <Box component="main" sx={{ flexGrow: 1, p: 2 }}>
-              <Toolbar />
-              <Router />
+          <GoogleAuthProvider>
+            <Box sx={{ display: " flex", height: "100vh" }}>
+              <Header />
+              <Sidebar />
+              <Box component="main" sx={{ flexGrow: 1, p: 2 }}>
+                <Toolbar />
+                <Router />
+              </Box>
             </Box>
-          </Box>
+          </GoogleAuthProvider>
         </AppThemeProvider>
       </BrowserRouter>
     </Provider>
